pyproject.toml
[project]
name = "apfa_locker_api"
version = "0.0.0-dev"
dependencies = [
    "fastapi[standard]",
    "aiohttp",
    "loguru",
    "bcrypt",
    "python-dotenv",
    "python-jose",
    "uvicorn[standard]>=0.30.6",
    ]


[project.optional-dependencies]
dev = [
    "ipython",
    "black",
    "autoflake",
    "flake8-absolute-import",
    "flake8-black",
    "flake8-bugbear",
    "flake8-cognitive-complexity",
    "flake8-eradicate",
    "flake8-isort",
    "flake8-fixme",
    "flake8-pep3101",
    "flake8-pie",
    "flake8-print",
    "flake8-printf-formatting",
    "flake8-pytest",
    "flake8-pytest-style",
    "flake8-simplify",
    "flake8-todo",
    "flake8-use-fstring",
    "flake8-variables-names",
    "flake8-walrus",
    "flake8-pyproject",

    "mypy",
    "types-freezegun",
    "types-Pillow",
    "types-python-jose",
    "pytest",
    "httpx",
]


[tool.flake8]
max-line-length = 160
inline-quotes = "\""
ignore = [
    "DJ05",  # URLs include() should set a namespace
    "E501",  # Line too long
    "E265",  # Block comments should have one space before the pound sign (#) and the comment itself
    "F811",  # Redefinition of unused name from line n
    "PT001",  # Use @pytest.fixture() over @pytest.fixture
    "SIM102",  # Use a single if-statement instead of nested if-statements
    "SIM113",  # Use enumerate instead of manually incrementing a counter
    "E203",  # whitespace before ':', disabled for black purposes https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices
    "B008",  # Do not perform function calls in argument defaults
    "PIE803",  # Disable prefer-logging-interpolation warning

]
exclude = [
    "static",
    "migrations",
    "frontend",
    ".git",
    "__pycache__",
    "src/app/",
]


[tool.isort]
profile = "google"
line_length = 160
extra_standard_library = ["pytest"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
use_parentheses = true
include_trailing_comma = true
multi_line_output = 3


[tool.black]
exclude = '''
/(
  | migrations
)/
'''
line_length = 160
